Aggressively clean up old EAP packets.
We track the State attribute via an internal rbtree.  When we
send a reply, we cache the REQUEST by State.  When we receive
a request, we look up the State, and delete the old REQUEST.

This lowers memory usage by ~40% in high-load EAP environments.

The problem is that "cleanup_delay" is too conservative.  When
we receive EAP packet N+1, we *know* we can delete EAP packets
1..N.  This is despite them being in "cleanup_delay" phase.

The next packet in the sequence is a positive acknowledgement from
the NAS that the previous packet was received.

The patch is in a separate branch because it is a new feature,
and should not go into v3.0.x